requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
python-dotenv
anthropic
sendgrid
twilio
